pyproject.toml
[project]
name = "skypulse-streaming-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bruin>=0.1.2",
    "dotenv>=0.9.9",
    "google-cloud-bigquery>=3.40.1",
    "google-cloud-bigquery-storage>=2.36.2",
    "google-cloud-storage>=3.10.0",
    "httpx>=0.27.0",
    "kafka-python>=2.3.0",
    "lxml>=6.0.2",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pyarrow>=23.0.1",
    "pydantic>=2.12.5",
    "setuptools>=68",
]
